requirements.txt
scrapy==2.11.2
beautifulsoup4==4.12.3
black==24.8.0
flake8==7.1.1
streamlit==1.39.0
plotly==5.24.1
matplotlib==3.9.2
SQLAlchemy==2.0.36
PyYAML==6.0.2
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0
google-auth-httplib2==0.2.0